LIBHYDRA-315. Refactored RepeatableComponents into individual files
Refactored React components in "RepeatableComponents" into individual
"RepeatablePlainLiteral" and "RepeatableTypedLiteral" components,
so that React Styleguidist would so them appropriately.

https://issues.umd.edu/browse/LIBHYDRA-315
