+++ app/views/react_components/react_components.html.erb
@@ -19,10 +19,12 @@
 
   <h3>PlainLiteral</h3>
 
+  <div>The following example is limited to 5 entries:<div>
+
   <div class="form-group">
     <%= react_component(
-          "RepeatableComponents.RepeatablePlainLiteral", {
-            maxValue: 10,
+          "RepeatablePlainLiteral", {
+            maxValues: 5,
             param_prefix: "repeatable_plain_literal",
             name: 'title',
             values: [
@@ -36,10 +38,11 @@
 
   <h3>TypedLiteral</h3>
 
+  <div>The following example has unlimited entries:</div>
+
   <div class="form-group">
     <%= react_component(
-          "RepeatableComponents.RepeatableTypedLiteral", {
-            maxValue: 10,
+          "RepeatableTypedLiteral", {
             param_prefix: "repeatable_typed_literal",
             defaultValue: { value: '', datatype: 'http://id.loc.gov/datatypes/edtf' },
             name: 'title',
